feat(search): moved in utils and implement removal
